requirements.txt
# PPT Agent - Dependencies

# LLM Clients
anthropic>=0.40.0
google-generativeai>=0.8.0

# PPT Generation
python-pptx>=1.0.0

# Utilities
python-dotenv>=1.0.0
